+++ emoncms-export.sh
@@ -66,7 +66,11 @@ for file in "${backup_location}/emoncms.sql" "${tar_filename}"
 do
     if [ -f "${file}" ]
     then
-        sudo rm "${file}"
+        if [ -f "/.dockerenv" ]; then
+            rm "${file}"
+        else
+            sudo rm "${file}"
+        fi
     fi
 done
 
@@ -92,7 +96,9 @@ fi
 #-----------------------------------------------------------------------------------------------
 
 # Disabled in @borphin commit?
-sudo systemctl stop feedwriter
+if [ ! -f "/.dockerenv" ]; then
+    sudo systemctl stop feedwriter
+fi
 
 # Get MYSQL authentication details from settings.php
 if [ -f "${script_location}/get_emoncms_mysql_auth.php" ]; then
@@ -101,7 +107,9 @@ if [ -f "${script_location}/get_emoncms_mysql_auth.php" ]; then
 else
     echo "Error: cannot read MYSQL authentication details from Emoncms $script_location/get_emoncms_mysql_auth.php php & settings.php"
     echo "$PWD"
-    sudo systemctl start feedwriter > /dev/null
+    if [ ! -f "/.dockerenv" ]; then
+        sudo systemctl start feedwriter > /dev/null
+    fi
     exit 1
 fi
 
@@ -111,12 +119,16 @@ if [ -n "$username" ]; then # if username string is not empty
     if [ $? -ne 0 ]; then
         echo "Error: failed to export mysql data"
         echo "emoncms export failed"
-        sudo systemctl start feedwriter > /dev/null
+        if [ ! -f "/.dockerenv" ]; then
+            sudo systemctl start feedwriter > /dev/null
+        fi
         exit 1
     fi
 else
     echo "Error: Cannot read MYSQL authentication details from Emoncms settings.php"
-    sudo systemctl start feedwriter > /dev/null
+    if [ ! -f "/.dockerenv" ]; then
+        sudo systemctl start feedwriter > /dev/null
+    fi
     exit 1
 fi
 
@@ -151,7 +163,9 @@ do
     fi
 done
 
-sudo systemctl start feedwriter > /dev/null
+if [ ! -f "/.dockerenv" ]; then
+    sudo systemctl start feedwriter > /dev/null
+fi
 
 # Compress backup
 echo "Compressing archive..."
